Move log formatter and handlers to arelle.logging module
