+++ arelle/Cntlr.py
@@ -18,8 +18,7 @@ import os
 import subprocess
 import sys
 import tempfile
-from collections import defaultdict
-from typing import Any, Mapping, TYPE_CHECKING, TextIO, cast
+from typing import Any, TYPE_CHECKING, cast
 
 import regex as re
 
@@ -27,6 +26,10 @@ from arelle import Locale, ModelManager, PackageManager, PluginManager, XbrlCons
 from arelle.BetaFeatures import BETA_FEATURES_AND_DESCRIPTIONS
 from arelle.SystemInfo import PlatformOS, getSystemWordSize, hasFileSystem, hasWebServer, isCGI, isGAE
 from arelle.WebCache import WebCache
+from arelle.logging.formatters import LogFormatter as _LogFormatter
+from arelle.logging.handlers.LogToBufferHandler import LogToBufferHandler
+from arelle.logging.handlers.LogToPrintHandler import LogToPrintHandler
+from arelle.logging.handlers.LogToXmlHandler import LogToXmlHandler
 from arelle.typing import TypeGetText
 from arelle.utils.PluginData import PluginData
 
@@ -38,6 +41,9 @@ if TYPE_CHECKING:
 osPrcs: Any = None
 LOG_TEXT_MAX_LENGTH = 32767
 cxFrozen = getattr(sys, 'frozen', False)
+
+logRefsFileLines = _LogFormatter.logRefsFileLines
+LogFormatter = _LogFormatter.LogFormatter
 
 
 def resourcesDir() -> str:
@@ -641,383 +647,3 @@ class Cntlr:
 
     def _clearPluginData(self) -> None:
         self.__pluginData.clear()
-
-
-def logRefsFileLines(refs: list[dict[str, Any]]) -> str:
-    fileLines = defaultdict(set)
-    for ref in refs:
-        href = ref.get("href")
-        if href:
-            fileLines[href.partition("#")[0]].add(ref.get("sourceLine") or 0)
-    return ", ".join(file + " " + ', '.join(str(line)
-                                            for line in sorted(lines, key=lambda l: l)
-                                            if line)
-                    for file, lines in sorted(fileLines.items()))
-
-class LogFormatter(logging.Formatter):
-    def __init__(self, fmt: str | None = None, datefmt: str | None = None) -> None:
-        super(LogFormatter, self).__init__(fmt, datefmt)
-
-    def fileLines(self, record: logging.LogRecord) -> str:
-        # provide a file parameter made up from refs entries
-        return logRefsFileLines(getattr(record, "refs", []))
-
-    def format(self, record: logging.LogRecord) -> str:
-        record.file = self.fileLines(record)
-        try:
-            formattedMessage = super(LogFormatter, self).format(record)
-        except (KeyError, TypeError, ValueError) as ex:
-            formattedMessage = "Message: "
-            if getattr(record, "messageCode", ""):
-                formattedMessage += "[{0}] ".format(getattr(record, "messageCode", ""))
-            if getattr(record, "msg", ""):
-                formattedMessage += record.msg + " "
-            if isinstance(record.args, dict) and 'error' in record.args: # args may be list or empty
-                formattedMessage += record.args['error']
-            formattedMessage += " \nMessage log error: " + str(ex)
-        if hasattr(record, "file"):
-            delattr(record, "file")
-        return formattedMessage
-
-class LogToPrintHandler(logging.Handler):
-    """
-    .. class:: LogToPrintHandler()
-
-    A log handler that emits log entries to standard out as they are logged.
-
-    CAUTION: Output is utf-8 encoded, which is fine for saving to files, but may not display correctly in terminal windows.
-
-    :param logOutput: 'logToStdErr' to cause log print to stderr instead of stdout
-    :type logOutput: str
-    """
-    logFile: str | TextIO | None
-
-    def __init__(self, logOutput: str) -> None:
-        super(LogToPrintHandler, self).__init__()
-        if logOutput == "logToStdErr":
-            self.logFile = sys.stderr
-        else:
-            self.logFile = None
-
-    def emit(self, logRecord: logging.LogRecord) -> None:
-        file = sys.stderr if self.logFile else None
-        logEntry = self.format(logRecord)
-        try:
-            print(logEntry, file=file)
-        except UnicodeEncodeError:
-            # extra parentheses in print to allow for 3-to-2 conversion
-            print((logEntry
-                   .encode(sys.stdout.encoding, 'backslashreplace')
-                   .decode(sys.stdout.encoding, 'strict')),
-                  file=file)
-
-class LogHandlerWithXml(logging.Handler):
-    logTextMaxLength: int
-    def __init__(self) -> None:
-        super(LogHandlerWithXml, self).__init__()
-
-    def recordToXml(self, logRec: logging.LogRecord) -> str:
-        def entityEncode(arg: Any, truncateAt: int = self.logTextMaxLength) -> str:  # be sure it's a string, vs int, etc, and encode &, <, ".
-            s = str(arg)
-            s = s if len(s) <= truncateAt else s[:truncateAt] + '...'
-            return s.replace("&","&amp;").replace("<","&lt;").replace('"','&quot;')
-
-        def ncNameEncode(arg: str) -> str:
-            s = []
-            for c in arg:
-                if c.isalnum() or c in ('.','-','_'):
-                    s.append(c)
-                else: # covers : and any other non-allowed character
-                    s.append('_') # change : into _ for xml correctness
-            return "".join(s)
-
-        def propElts(properties: list[tuple[Any, Any, Any]], indent: str, truncateAt: int = 128) -> str:
-            nestedIndent = indent + ' '
-            return indent.join('<property name="{0}" value="{1}"{2}>'.format(
-                                    entityEncode(p[0]),
-                                    entityEncode(p[1], truncateAt=truncateAt),
-                                    '/' if len(p) == 2
-                                    else '>' + nestedIndent + propElts(p[2],nestedIndent) + indent + '</property')
-                                for p in properties
-                                if 2 <= len(p) <= 3)
-
-        msg = self.format(logRec)
-        if logRec.args and isinstance(logRec.args, Mapping):
-            args = "".join([' {0}="{1}"'.format(ncNameEncode(n), entityEncode(v,
-                                                truncateAt=(4096000 if n in ("json",) else 128)))
-                            for n, v in logRec.args.items()])
-        else:
-            args = ""
-        refs = "\n ".join('\n <ref href="{0}"{1}{2}{3}>'.format(
-                        entityEncode(ref["href"]),
-                        ' sourceLine="{0}"'.format(ref["sourceLine"]) if "sourceLine" in ref else '',
-                        ''.join(' {}="{}"'.format(ncNameEncode(k),entityEncode(v))
-                                                  for k,v in ref["customAttributes"].items())
-                             if 'customAttributes' in ref else '',
-                        (">\n  " + propElts(ref["properties"],"\n  ", truncateAt=self.logTextMaxLength) + "\n </ref" )
-                                   if ("properties" in ref) else '/')
-                       for ref in getattr(logRec, "refs", []))
-        return ('<entry code="{0}" level="{1}">'
-                '\n <message{2}>{3}</message>{4}'
-                '</entry>\n'.format(getattr(logRec, "messageCode", ""),
-                                    logRec.levelname.lower(),
-                                    args,
-                                    entityEncode(msg),
-                                    refs))
-    def recordToJson(self, logRec: logging.LogRecord) -> dict[str, Any]:
-        message = { "text": self.format(logRec) }
-        if logRec.args and isinstance(logRec.args, Mapping):
-            for n, v in logRec.args.items():
-                message[n] = str(v)
-        return {"code": getattr(logRec, "messageCode", ""),
-                "level": logRec.levelname.lower(),
-                "refs": getattr(logRec, "refs", []),
-                "message": message}
-
-    def recordToHtml(self, logRec: logging.LogRecord) -> str:
-        record = ["<tr>"]
-        record.append(f"<td>{getattr(logRec, 'messageCode', '')}</td>")
-        record.append(f"<td>{logRec.levelname.lower()}</td>")
-        record.append(f"<td>{self.format(logRec)}</td>")
-        record.append("</tr>")
-        return "\n".join(record)
-
-class LogToXmlHandler(LogHandlerWithXml):
-    """
-    .. class:: LogToXmlHandler(filename)
-
-    A log handler that writes log entries to named XML file (utf-8 encoded) upon closing the application.
-    """
-    logRecordBuffer: list[logging.LogRecord]
-    filename: str | None
-    filemode: str
-    htmlTitle: str = "Arelle Message Log" # may be customized in plugin startup
-
-    def __init__(self, filename: str | None = None, mode: str = 'w') -> None:
-        super(LogToXmlHandler, self).__init__()
-        self.filename = filename # may be none if buffer is retrieved by get methods below and not written anywhere
-        self.logRecordBuffer = []
-        self.filemode = mode
-
-    def flush(self) -> None:
-        # Note to developers: breakpoints in this method don't work, please debug with print statements
-        securityIsActive = securityHasWritten = False
-        for pluginMethod in PluginManager.pluginClassMethods("Security.Crypt.IsActive"):
-            securityIsActive = pluginMethod(self) # must be active for the cntlr object to effect log writing
-        if self.filename == "logToStdOut.xml":
-            print('<?xml version="1.0" encoding="utf-8"?>')
-            print('<log>')
-            for logRec in self.logRecordBuffer:
-                logRecXml = self.recordToXml(logRec)
-                try:
-                    print(logRecXml)
-                except UnicodeEncodeError:
-                    # extra parentheses in print to allow for 3-to-2 conversion
-                    print((logRecXml
-                           .encode(sys.stdout.encoding, 'backslashreplace')
-                           .decode(sys.stdout.encoding, 'strict')))
-            print('</log>')
-        elif self.filename is not None:
-            if self.filename.endswith(".xml"):
-                # print ("filename=" + self.filename)
-                if securityIsActive:
-                    for pluginMethod in PluginManager.pluginClassMethods("Security.Crypt.Write"):
-                        securityHasWritten = pluginMethod(self, self.filename,
-                            '<?xml version="1.0" encoding="utf-8"?>\n<log>\n' +
-                            ''.join(self.recordToXml(logRec) for logRec in self.logRecordBuffer) +
-                            '</log>\n')
-                if not securityHasWritten:
-                    with open(self.filename, self.filemode, encoding='utf-8') as fh:
-                        fh.write('<?xml version="1.0" encoding="utf-8"?>\n<log>\n')
-                        for logRec in self.logRecordBuffer:
-                            fh.write(self.recordToXml(logRec))
-                        fh.write('</log>\n')
-            elif self.filename.endswith(".json"):
-                if securityIsActive:
-                    for pluginMethod in PluginManager.pluginClassMethods("Security.Crypt.Write"):
-                        securityHasWritten = pluginMethod(self, self.filename, self.getJson())
-                if not securityHasWritten:
-                    with open(self.filename, self.filemode, encoding='utf-8') as fh:
-                        fh.write(self.getJson())
-            elif self.filename.endswith(".html"):
-                if securityIsActive:
-                    for pluginMethod in PluginManager.pluginClassMethods("Security.Crypt.Write"):
-                        securityHasWritten = pluginMethod(self, self.filename, self.getHtml())
-                if not securityHasWritten:
-                    with open(self.filename, self.filemode, encoding='utf-8') as fh:
-                        fh.write(self.getHtml())
-            elif self.filename in ("logToPrint", "logToStdErr"):
-                _file = sys.stderr if self.filename == "logToStdErr" else None
-                for logRec in self.logRecordBuffer:
-                    logEntry = self.format(logRec)
-                    try:
-                        print(logEntry, file=_file)
-                    except UnicodeEncodeError:
-                        # extra parentheses in print to allow for 3-to-2 conversion
-                        print((logEntry
-                               .encode(sys.stdout.encoding, 'backslashreplace')
-                               .decode(sys.stdout.encoding, 'strict')),
-                              file=_file)
-            else:
-                if securityIsActive:
-                    for pluginMethod in PluginManager.pluginClassMethods("Security.Crypt.Write"):
-                        securityHasWritten = pluginMethod(self, self.filename,
-                            ''.join(self.format(logRec) + "\n" for logRec in self.logRecordBuffer))
-                if not securityHasWritten:
-                    with open(self.filename, self.filemode, encoding='utf-8') as fh:
-                        for logRec in self.logRecordBuffer:
-                            fh.write(self.format(logRec) + "\n")
-        self.clearLogBuffer()
-
-    def clearLogBuffer(self) -> None:
-        del self.logRecordBuffer[:]
-
-    def getXml(self, clearLogBuffer: bool = True, includeDeclaration: bool = True) -> str:
-        """
-        Returns an XML document (as a string) representing the messages in the log buffer.
-        Optionally clears the log buffer afterwards.
-        Optionally includes XML declaration.
-        :return: XML string of messages in the log buffer.
-        """
-        xml = []
-        if includeDeclaration:
-            xml.append('<?xml version="1.0" encoding="utf-8"?>\n')
-        xml.append('<log>')
-        for logRec in self.logRecordBuffer:
-            xml.append(self.recordToXml(logRec))
-        xml.append('</log>')
-        if clearLogBuffer:
-            self.clearLogBuffer()
-        return '\n'.join(xml)
-
-    def getJson(self, clearLogBuffer: bool = True) -> str:
-        """Returns an JSON string representing the messages in the log buffer, and clears the buffer.
-
-        :returns: str -- json representation of messages in the log buffer
-        """
-        entries = []
-        for logRec in self.logRecordBuffer:
-            entries.append(self.recordToJson(logRec))
-        if clearLogBuffer:
-            self.clearLogBuffer()
-        return json.dumps( {"log": entries}, ensure_ascii=False, indent=1, default=str )
-
-    def getHtml(self, clearLogBuffer: bool = True) -> str:
-        """Returns an HTML string representing the messages in the log buffer, and clears the buffer.
-
-        :returns: str -- HTML representation of messages in the log buffer
-        """
-        html = ["""<!doctype html>
-        <html>
-        <head>
-            <title>{0}</title>
-            <style>
-                table {{
-                    border: 1px solid black;
-                    border-spacing: 3px;
-                    table-layout: fixed;
-                    width: 100%;
-                }}
-                th, td {{
-                    padding: 5px;
-                    word-wrap: break-word;
-                }}
-                th {{
-                    background-color: #bcf1fd;
-                }}
-                td {{
-                    background-color: #f4f7f7;
-                }}
-                td:last-child, th:last-child {{
-                    width: 80%;
-                }}
-            </style>
-        </head>
-        <body>
-        <table>
-            <thead>
-                <tr>
-                    <th>Code</th>
-                    <th>Level</th>
-                    <th style="">Message</th>
-                </tr>
-            </thead>
-            <tbody>""".format(
-                self.htmlTitle)
-        ]
-        if self.logRecordBuffer:
-            for logRec in self.logRecordBuffer:
-                if all(p(logRec) for p in PluginManager.pluginClassMethods("Cntlr.Log.RecFilter.Html")):
-                    html.append(self.recordToHtml(logRec))
-            if clearLogBuffer:
-                self.clearLogBuffer()
-            html.append("</tbody>\n</table>\n</body>\n</html>\n")
-        if len(html) < 3: # no entries were added to log. Display no log errors message
-            html = ["""<!doctype html>
-            <html>
-            <head>
-                <title>{0}</title>
-                <style>
-                    div {{
-                        background-color: #f4f7f7;
-                        position: relative;
-                        top: 33.33%;
-                        margin: -10px auto;
-                        width: 25vw;
-                        height: 25vh;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        text-align: center;
-                    }}
-                    body {{
-                        background-color: #bcf1fd;
-                        height: 100vh;
-                    }}
-                </style>
-            </head>
-            <body>
-                <div>
-                    <h1>No log errors to display</h1>
-                </div>
-            </body>
-            </html>
-            """.format(
-                self.htmlTitle)
-            ]
-        return '\n'.join(html)
-
-    def getLines(self, clearLogBuffer: bool = True) -> list[str]:
-        """Returns a list of the message strings in the log buffer, and clears the buffer.
-
-        :returns: [str] -- list of strings representing messages corresponding to log buffer entries
-        """
-        lines = [self.format(logRec) for logRec in self.logRecordBuffer]
-        if clearLogBuffer:
-            self.clearLogBuffer()
-        return lines
-
-    def getText(self, separator: str = '\n', clearLogBuffer: bool = True) -> str:
-        """Returns a string of the lines in the log buffer, separated by newline or provided separator.
-
-        :param separator: Line separator (default is platform os newline character)
-        :type separator: str
-        :returns: str -- joined lines of the log buffer.
-        """
-        return separator.join(self.getLines(clearLogBuffer=clearLogBuffer))
-
-    def emit(self, logRecord: logging.LogRecord) -> None:
-        self.logRecordBuffer.append(logRecord)
-
-class LogToBufferHandler(LogToXmlHandler):
-    """
-    .. class:: LogToBufferHandler()
-
-    A log handler that writes log entries to a memory buffer for later retrieval (to a string) in XML, JSON, or text lines,
-    usually for return to a web service or web page call.
-    """
-    def __init__(self) -> None:
-        super(LogToBufferHandler, self).__init__()
-
-    def flush(self) -> None:
-        pass # do nothing -- overrides LogToXmlHandler's flush
